work for the refactoring/circular-dep-and-loosely-coupling
