+++ packages/survey-creator-core/src/drop-to-enum.ts
@@ -1,0 +1,7 @@
+export enum DropTo {
+  Inside = "inside",
+  Top = "top",
+  Bottom = "bottom",
+  Right = "right",
+  Left = "left"
+}
