+++ packages/survey-creator-core/tests/dragdrop-elements.tests.ts
@@ -1,5 +1,6 @@
 import { Question, QuestionTextModel, SurveyModel } from "survey-core";
-import { DragDropSurveyElements, calculateDragOverLocation, calculateIsEdge, calculateIsSide, DropTo } from "../src/dragdrop-survey-elements";
+import { DragDropSurveyElements, calculateDragOverLocation, calculateIsEdge, calculateIsSide } from "../src/dragdrop-survey-elements";
+import { DropTo } from "../src/drop-to-enum";
 import { CreatorTester } from "./creator-tester";
 import { QuestionAdornerViewModel } from "../src/components/question";
 import { PageAdorner } from "../src/components/page";
