+++ packages/survey-creator-core/tests/dragdrop.tests.ts
@@ -1,5 +1,6 @@
 import { SurveyModel, QuestionTextModel } from "survey-core";
-import { DragDropSurveyElements, DropTo } from "../src/dragdrop-survey-elements";
+import { DragDropSurveyElements } from "../src/dragdrop-survey-elements";
+import { DropTo } from "../src/drop-to-enum";
 import { CreatorTester } from "./creator-tester";
 import { QuestionAdornerViewModel } from "../src/components/question";
